fix: two full-gate clippy errors at HEAD (checklist cast + example match-to-?)
The full gate (full_gate.yml, last run 2026-09-03 15:31 @11a5b84c) reds on 2
error diagnostics at HEAD; the Monday 09-08 scheduled rot check would red on
them:
- crates/katgpt-claim/src/claim_rubric/checklist.rs:304: .len() as usize is a
  same-type cast, denied by the gate's -D clippy::unnecessary_cast.
- examples/dec_terrain_quality_bench.rs:189: match best { Some(next) => {...}
  None => return None } -> let next = best?; denied by -D clippy::question_mark.
Both validated with the gate's deny list scoped to the two targets (clippy
clean) + section_items tests pass. No behavior change.

+++ crates/katgpt-claim/src/claim_rubric/checklist.rs
@@ -301,7 +301,7 @@ mod tests {
     #[test]
     fn section_items_match_constants() {
         assert_eq!(section_items(ChecklistSection::S1).len(), S1_ITEMS.len());
-        assert_eq!(section_items(ChecklistSection::S4).len() as usize, 5);
+        assert_eq!(section_items(ChecklistSection::S4).len(), 5);
     }
 
     #[test]
